+++ Changelog.md
@@ -12,3 +12,5 @@
 * Updated viewmodel and camera movement (Inspired by cancelled game, Boreal Alyph)
 ### 12
 * Removed most of the telemetry (remainders left in `/togl*/glmgrbasics.cpp`)
+* /public/mathlib/polyhedron.cpp: CreateDumpDirectory now supports Linux
+* Added .sh format for scripts in /linux/